move FieldValue and ComputedValue to common module.
