+++ core/connectors/sdk/src/transforms/add_fields.rs
@@ -16,45 +16,15 @@
  * under the License.
  */
 
-use super::{Transform, TransformType};
+use super::{FieldValue, Transform, TransformType};
 use crate::{DecodedMessage, Error, Payload, TopicMetadata};
 use serde::{Deserialize, Serialize};
-use simd_json::OwnedValue;
-use strum_macros::{Display, IntoStaticStr};
 
 /// A field to be added to messages
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Field {
     pub key: String,
     pub value: FieldValue,
-}
-
-/// The value of a field, either static or computed at runtime
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum FieldValue {
-    Static(OwnedValue),
-    Computed(ComputedValue),
-}
-
-/// Types of computed values that can be generated at runtime
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Serialize, Deserialize, Display, IntoStaticStr)]
-#[serde(rename_all = "snake_case")]
-pub enum ComputedValue {
-    #[strum(to_string = "date_time")]
-    DateTime,
-    #[strum(to_string = "timestamp_nanos")]
-    TimestampNanos,
-    #[strum(to_string = "timestamp_micros")]
-    TimestampMicros,
-    #[strum(to_string = "timestamp_millis")]
-    TimestampMillis,
-    #[strum(to_string = "timestamp_seconds")]
-    TimestampSeconds,
-    #[strum(to_string = "uuid_v4")]
-    UuidV4,
-    #[strum(to_string = "uuid_v7")]
-    UuidV7,
 }
 
 /// Configuration for the AddFields transform
