+++ core/connectors/sdk/src/transforms/json/add_fields.rs
@@ -16,13 +16,11 @@
  * under the License.
  */
 
-use chrono::Utc;
-use simd_json::OwnedValue;
-
 use crate::{
     DecodedMessage, Error, Payload, TopicMetadata,
-    transforms::add_fields::{AddFields, ComputedValue, FieldValue},
+    transforms::{AddFields, FieldValue, compute_value},
 };
+use simd_json::OwnedValue;
 
 impl AddFields {
     pub(crate) fn transform_json(
@@ -46,19 +44,6 @@ impl AddFields {
     }
 }
 
-fn compute_value(kind: &ComputedValue) -> OwnedValue {
-    let now = Utc::now();
-    match kind {
-        ComputedValue::DateTime => now.to_rfc3339().into(),
-        ComputedValue::TimestampNanos => now.timestamp_nanos_opt().unwrap().into(),
-        ComputedValue::TimestampMicros => now.timestamp_micros().into(),
-        ComputedValue::TimestampMillis => now.timestamp_millis().into(),
-        ComputedValue::TimestampSeconds => now.timestamp().into(),
-        ComputedValue::UuidV4 => uuid::Uuid::new_v4().to_string().into(),
-        ComputedValue::UuidV7 => uuid::Uuid::now_v7().to_string().into(),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -67,8 +52,8 @@ mod tests {
         create_test_topic_metadata, extract_json_object,
     };
     use crate::transforms::{
-        Transform,
-        add_fields::{AddFields, ComputedValue, Field, FieldValue},
+        ComputedValue, FieldValue, Transform,
+        add_fields::{AddFields, Field},
     };
     use simd_json::OwnedValue;
